requirements.txt
Django==1.8.3
DjangoRestless==0.0.10